Adds LLVM's LLD linker and ThinLTO as options when building(p=x11)
When building with the SCons command "use_llvm=yes" and "use_lld=yes",
this will make Clang use LLD as the linker. LLD is normally only called
by the compiler driver (LLVM/Clang in this case) or else the system
defaults to GCC's default ld, even when Clang is used as a compiler. But
since not every Clang user also uses LLD I've set it to false per
default.

In order to also use ThinLTO one would add this when calling SCons:
"use_llvm=yes use_lld=yes use_thinlto=yes"

Support is added only to Linux and in combination with clang, so
use_llvm=yes is mandatory to be able to use LLD linker. For being able
to use ThinLTO for now also I only added support for the combination
Clang and LDD so for ThinLTO you need also use_llvm and use_lld set to
yes when calling SCons.
For people using other compilers and linkers and that use either of the
two (LDD and ThinLTO) or both, they can still pass the flags directly
when calling SCons as usual when setting compiler/linker flags/options.
This commit makes it just a little less typing and fits into how we use
SCons.

Hint: "use_lld" and "use_thinlto" are both set to false per default.